+++ src/lib/provable/group.ts
@@ -11,6 +11,7 @@ import {
 import { Provable } from './provable.js';
 import { Bool } from './bool.js';
 import { assert } from '../util/assert.js';
+import { add } from './gadgets/native-curve.js';
 
 export { Group };
 
@@ -106,7 +107,7 @@ class Group {
         return fromProjective(g_proj);
       }
     } else {
-      let { result, isInfinity } = addBase(this, g);
+      let { result, isInfinity } = add(this, g);
       // similarly to the constant implementation, we check if either operand is zero
       // and the implementation above (original OCaml implementation) returns something wild -> g + 0 != g where it should be g + 0 = g
       let gIsZero = g.isZero();
@@ -139,7 +140,7 @@ class Group {
       assert(infinity === false, 'Group.addNonzero(): Result is zero');
       return fromAffine({ x, y, infinity });
     }
-    let { result, isInfinity } = addBase(this, g2);
+    let { result, isInfinity } = add(this, g2);
 
     if (allowZeroOutput) {
       return Provable.if(isInfinity, Group.zero, new Group(result));
@@ -340,63 +341,6 @@ class Group {
       fields: [x.x, x.y],
     };
   }
-}
-
-type GroupBase = { x: Field; y: Field };
-
-/**
- * Wraps the `EC_add` gate to perform complete addition of two non-zero curve points.
- */
-function addBase(g: GroupBase, h: GroupBase) {
-  const { x: x1, y: y1 } = g;
-  const { x: x2, y: y2 } = h;
-
-  let zero = new Field(0);
-
-  let same_x = Provable.witness(Field, () => x1.equals(x2).toField());
-
-  let inf = Provable.witness(Bool, () =>
-    x1.equals(x2).and(y1.equals(y2).not())
-  );
-
-  let inf_z = Provable.witness(Field, () => {
-    if (y1.equals(y2).toBoolean()) return zero;
-    else if (x1.equals(x2).toBoolean()) return y2.sub(y1).inv();
-    else return zero;
-  });
-
-  let x21_inv = Provable.witness(Field, () => {
-    if (x1.equals(x2).toBoolean()) return zero;
-    else return x2.sub(x1).inv();
-  });
-
-  let s = Provable.witness(Field, () => {
-    if (x1.equals(x2).toBoolean()) {
-      let x1_squared = x1.square();
-      return x1_squared.add(x1_squared).add(x1_squared).div(y1.add(y1));
-    } else return y2.sub(y1).div(x2.sub(x1));
-  });
-
-  let x3 = Provable.witness(Field, () => {
-    return s.square().sub(x1.add(x2));
-  });
-
-  let y3 = Provable.witness(Field, () => {
-    return s.mul(x1.sub(x3)).sub(y1);
-  });
-
-  Snarky.gates.ecAdd(
-    toTuple(Group.from(x1.seal(), y1.seal())),
-    toTuple(Group.from(x2.seal(), y2.seal())),
-    toTuple(Group.from(x3, y3)),
-    inf.toField().value,
-    same_x.value,
-    s.value,
-    inf_z.value,
-    x21_inv.value
-  );
-
-  return { result: { x: x3, y: y3 }, isInfinity: inf };
 }
 
 // internal helpers
